Feature/add kafka

+++ src/OzonEdu.StockApi.HttpModels/DeliveryRequestViewModel.cs
@@ -5,9 +5,9 @@ namespace OzonEdu.StockApi.HttpModels
 {
     public record DeliveryRequestViewModel
     {
-        public int Id { get; init; }
+        public long Id { get; init; }
 
-        public int DeliveryRequestId { get; init; }
+        public long DeliveryRequestId { get; init; }
 
         public DeliveryRequestStatus RequestStatus { get; init; }
