+++ src/OzonEdu.StockApi.Infrastructure/Commands/CreateDeliveryRequest/CreateDeliveryRequestCommand.cs
@@ -4,7 +4,7 @@ using OzonEdu.StockApi.Infrastructure.Models;
 
 namespace OzonEdu.StockApi.Infrastructure.Commands.CreateDeliveryRequest
 {
-    public class CreateDeliveryRequestCommand : IRequest<int>, IItemsModel<DeliveryRequestDto>
+    public class CreateDeliveryRequestCommand : IRequest<long>, IItemsModel<DeliveryRequestDto>
     {
         public IReadOnlyList<DeliveryRequestDto> Items { get; set; }
     }
